+++ lnmp-docker.ps1
@@ -110,7 +110,7 @@ if ($args[0] -eq "env-file") {
   exit
 }
 
-if (!(Test-Path $PSScriptRoot\cli\khs1994-robot.enc )) {
+if (!(Test-Path $PSScriptRoot\scripts\cli\khs1994-robot.enc )) {
   Write-Host "lnmp-docker.ps1 not in lnmp ROOT PATH" -ForegroundColor Red
 
   Write-Host "Please remove $((get-command lnmp-docker).Source)" -ForegroundColor Red
@@ -122,7 +122,7 @@ printInfo "Load env file [ $LNMP_ENV_FILE ] and [ $LNMP_ENV_FILE_PS1 ]"
 
 . "$PSScriptRoot/.env.example.ps1"
 
-. "$PSScriptRoot/cli/.env.ps1"
+. "$PSScriptRoot/scripts/cli/.env.ps1"
 
 if (Test-Path "$PSScriptRoot/$LNMP_ENV_FILE_PS1") {
   . "$PSScriptRoot/$LNMP_ENV_FILE_PS1"
@@ -572,29 +572,9 @@ Function Get-ComposeOptions($compose_files) {
 
   $COMPOSE_FILE_ARRAY += "docker-workspace.yml"
 
-  if ($env:USE_WSL2_DOCKER_COMPOSE -eq '1') {
-    $COMPOSE_FILE_ARRAY_FULL_PATH = @()
-
-    foreach ($item in $COMPOSE_FILE_ARRAY) {
-      $COMPOSE_FILE_ARRAY_FULL_PATH += (Get-Item $item).FullName
-    }
-
-    $env:COMPOSE_FILE = $COMPOSE_FILE_ARRAY_FULL_PATH -join ';'
-
-    $env:WSLENV = 'COMPOSE_FILE/ul'
-
-    #@debug
-    printInfo $(wsl -d ${WSL2_DIST} -- sh -c 'echo Load compose files: $COMPOSE_FILE')
-  }
-  else {
-    $env:COMPOSE_FILE = $COMPOSE_FILE_ARRAY -join ';'
-  }
+  $env:COMPOSE_FILE = $COMPOSE_FILE_ARRAY -join ';'
 
   # write-host $env:COMPOSE_FILE
-
-  if ($env:USE_WSL2_DOCKER_COMPOSE -eq '1') {
-    return $options
-  }
 
   return $options.split(' ')
 }
@@ -615,26 +595,6 @@ function Copy-PCIT() {
   docker pull pcit/pcit:frontend
 
   docker run -it --rm -v ${APP_ROOT}/.pcit/public:/var/www/pcit/public pcit/pcit:frontend
-}
-
-function convert_args_to_string_if_use_wsl2() {
-  if ($env:USE_WSL2_DOCKER_COMPOSE -eq '0') {
-    return $args
-  }
-
-  $string = ""
-
-  foreach ($item in $args) {
-    if ($item.getType().Name -eq 'Object[]' ) {
-      foreach ($i in $item) {
-        $string += " $i"
-      }
-      continue
-    }
-    $string += " $item"
-  }
-
-  return $string
 }
 
 function Edit-Hosts() {
@@ -679,59 +639,30 @@ function Get-Env($ENV_NAME, $ENV_FILE, $ENV_DEFAULT) {
 # APP_ROOT
 $APP_ROOT = Get-Env 'APP_ROOT' $LNMP_ENV_FILE './app'
 
-$env:USE_WSL2_DOCKER_COMPOSE = '0'
-$env:USE_WSL2_BUT_DOCKER_NOT_RUNNING = '0'
+$env:USE_WSL2_BUT_DOCKER_NOT_RUNNING = '3'
 
-# 判断项目是否存储在 WSL2
+if ($APP_ROOT.Substring(0, 7) -eq '\\wsl$\') {
+  $env:USE_WSL2_BUT_DOCKER_NOT_RUNNING = '0'
 
-function docker-compose() {
-  foreach ($item in $args) {
-    if (!($item)) { continue }
-    if ($item.getType().Name -eq 'Object[]' ) {
-      foreach ($i in $item) {
-        $string += " $i"
-      }
-      continue
-    }
-    $string += " $item"
+  $WSL2_DIST = $APP_ROOT.Split('\')[3]
+
+  if ($APP_ROOT.length -le (8 + $WSL2_DIST.length)) {
+    printError "please check APP_ROOT value in .env file"
+
+    exit 1
   }
 
-  printInfo "Load compose file: $env:COMPOSE_FILE"
-  printInfo "Exec:              $ docker compose $string"
+  $WSL2_DIST_PATH = '/' + $APP_ROOT.Substring(8 + $WSL2_DIST.length)
 
-  Invoke-Expression "docker compose $string"
-}
+  if ($WSL2_DIST_PATH.Length -eq 1) {
+    printError "please check APP_ROOT value in .env file"
 
-if ($APP_ROOT.Substring(0, 1) -eq '/' -and $WSL2_DIST) {
-  $env:USE_WSL2_DOCKER_COMPOSE = '1'
+    exit 1
+  }
 
   if (!(Test-Path \\wsl$\$WSL2_DIST/mnt/wsl/docker-desktop/cli-tools/usr/bin/docker)) {
     $env:USE_WSL2_BUT_DOCKER_NOT_RUNNING = '1'
   }
-
-  function docker-compose() {
-    #@debug
-    printInfo "Exec docker compose command in WSL2 [ ${WSL2_DIST} ]"
-
-    $COMPOSE_BIN = "/mnt/wsl/docker-desktop/cli-tools/usr/bin/docker"
-
-    $cmd = convert_args_to_string_if_use_wsl2 $args
-
-    wsl -d $WSL2_DIST -- sh -c "command -v $COMPOSE_BIN > /dev/null || exit 1"
-
-    if ($?) {
-      wsl -d $WSL2_DIST -- sh -xc "$COMPOSE_BIN compose $cmd"
-    }
-    else {
-      printWarning "Docker Desktop not running"
-    }
-  }
-}
-
-if ($APP_ROOT.Substring(0, 1) -eq '/' -and !($WSL2_DIST)) {
-  printError "APP_ROOT $APP_ROOT start with '/', but `$WSL2_DIST not set in .env.ps1"
-
-  exit 1
 }
 
 # APP_ENV
@@ -739,12 +670,12 @@ if ($APP_ROOT.Substring(0, 1) -eq '/' -and !($WSL2_DIST)) {
 $APP_ENV = Get-Env -ENV_NAME 'APP_ENV' -ENV_FILE $LNMP_ENV_FILE -ENV_DEFAULT 'development'
 
 # cd LNMP_ROOT
-if (!(Test-Path cli/khs1994-robot.enc )) {
+if (!(Test-Path scripts/cli/khs1994-robot.enc )) {
   # 在项目目录外
   printInfo "Use LNMP CLI in $PWD"
   cd $PSScriptRoot
-  if ($APP_ROOT.Substring(0, 1) -eq '/' ) {
-    printInfo "APP_ROOT is WSL2 [ $WSL2_DIST ] PATH $APP_ROOT"
+  if ($APP_ROOT.Substring(0, 7) -eq '\\wsl$\') {
+    printInfo "APP_ROOT is WSL2 [ $WSL2_DIST ] PATH $WSL2_DIST_PATH"
   }
   else {
     # cd $PSScriptRoot
@@ -759,8 +690,8 @@ if (!(Test-Path cli/khs1994-robot.enc )) {
 }
 else {
   printInfo "Use LNMP CLI in LNMP Root $pwd"
-  if ($APP_ROOT.Substring(0, 1) -eq '/' ) {
-    printInfo "APP_ROOT is WSL2 [ $WSL2_DIST ] PATH $APP_ROOT"
+  if ($APP_ROOT.Substring(0, 7) -eq '\\wsl$\') {
+    printInfo "APP_ROOT is WSL2 [ $WSL2_DIST ] PATH $WSL2_DIST_PATH"
   }
   else {
     if (!(Test-Path $APP_ROOT)) {
@@ -826,6 +757,12 @@ if (!($?)) {
   exit 1
 }
 
+if (!((docker compose version).split( )[-1].trim('v') -ge $REQUIRE_DOCKER_COMPOSE_VERSION)) {
+  printError "your compose version is not ge v$REQUIRE_DOCKER_COMPOSE_VERSION"
+
+  exit 1
+}
+
 printInfo $(docker compose version)
 
 if ($args.Count -eq 0) {
@@ -875,8 +812,6 @@ switch -regex ($command) {
   }
 
   "^build$" {
-    $env:USE_WSL2_DOCKER_COMPOSE = '0'
-
     if ($other) {
       $services = $other
     }
@@ -893,8 +828,6 @@ switch -regex ($command) {
   }
 
   push {
-    $env:USE_WSL2_DOCKER_COMPOSE = '0'
-
     if ($other) {
       $services = $other
     }
@@ -922,7 +855,7 @@ switch -regex ($command) {
     $options = Get-ComposeOptions "docker-lnmp.yml", `
       "docker-lnmp.override.yml"
 
-    & { docker-compose ${LNMP_COMPOSE_GLOBAL_OPTIONS} $options config $other }
+    docker compose ${LNMP_COMPOSE_GLOBAL_OPTIONS} $options config $other
   }
 
   checkout {
@@ -951,15 +884,14 @@ switch -regex ($command) {
     #@custom
     __lnmp_custom_pre_up $services
 
-    & { docker-compose ${LNMP_COMPOSE_GLOBAL_OPTIONS} $options up --no-build -d $services }
+    docker compose ${LNMP_COMPOSE_GLOBAL_OPTIONS} $options config > compose-up.yaml
+    docker compose ${LNMP_COMPOSE_GLOBAL_OPTIONS} $options up --no-build -d $services
 
     #@custom
     __lnmp_custom_post_up $services
   }
 
   "^pull$" {
-    $env:USE_WSL2_DOCKER_COMPOSE = '0'
-
     if ($other) {
       $services = $other
     }
@@ -980,8 +912,6 @@ switch -regex ($command) {
   }
 
   "^down$" {
-    $env:USE_WSL2_DOCKER_COMPOSE = '0'
-
     $options = Get-ComposeOptions "docker-lnmp.yml", `
       "docker-lnmp.override.yml"
 
@@ -1026,8 +956,6 @@ switch -regex ($command) {
   }
 
   restart {
-    $env:USE_WSL2_DOCKER_COMPOSE = '0'
-
     $options = Get-ComposeOptions "docker-lnmp.yml", `
       "docker-lnmp.override.yml"
 
@@ -1330,7 +1258,7 @@ XXX
     $options = Get-ComposeOptions "docker-lnmp.yml", `
       "docker-lnmp.override.yml"
 
-    & { docker-compose ${LNMP_COMPOSE_GLOBAL_OPTIONS} $options up -d ${LNMP_SERVICES} pcit }
+    docker compose ${LNMP_COMPOSE_GLOBAL_OPTIONS} $options up -d ${LNMP_SERVICES} pcit
   }
 
   daemon-socket {
@@ -1357,7 +1285,7 @@ XXX
     if ((Test-Path $EXEC_CMD_DIR/.devcontainer) -And (Test-Path $EXEC_CMD_DIR/docker-workspace.yml)) {
       printInfo "Exec composer command in [ vscode remote ] or [ PhpStorm ] project folder"
       cd $EXEC_CMD_DIR
-      docker-compose ${LNMP_COMPOSE_GLOBAL_OPTIONS} -f docker-workspace.yml run --rm composer $args
+      docker compose ${LNMP_COMPOSE_GLOBAL_OPTIONS} -f docker-workspace.yml run --rm composer $args
 
       exit
     }
@@ -1378,7 +1306,7 @@ Example: ./lnmp-docker composer /app/demo install
     $options = Get-ComposeOptions "docker-lnmp.yml", `
       "docker-lnmp.override.yml"
 
-    & { docker-compose ${LNMP_COMPOSE_GLOBAL_OPTIONS} $options run -w $WORKING_DIR --rm composer $COMPOSER_COMMAND }
+    docker compose ${LNMP_COMPOSE_GLOBAL_OPTIONS} $options run -w $WORKING_DIR --rm composer $COMPOSER_COMMAND
   }
 
   hosts {
@@ -1396,7 +1324,7 @@ Example: ./lnmp-docker composer /app/demo install
   "^code-run$" {
     cd $EXEC_CMD_DIR
 
-    docker-compose -f docker-workspace.yml run --rm $other
+    docker compose -f docker-workspace.yml run --rm $other
   }
 
   "^code-exec$" {
@@ -1406,6 +1334,13 @@ Example: ./lnmp-docker composer /app/demo install
   }
 
   "mount" {
+    if ($env:USE_WSL2_BUT_DOCKER_NOT_RUNNING -eq '0') {
+      printError "Docker is running"
+      printError "Please run [ ./lnmp-docker mount ] before docker start"
+
+      exit 1
+    }
+
     if (!$WSL2_DIST) {
       $WSL2_DIST = 'ubuntu'
     }
@@ -1423,10 +1358,10 @@ Example: ./lnmp-docker composer /app/demo install
     }
 
     printInfo "try mount physical disk to WSL2 [ $WSL2_DIST ]"
-    wsl -d $WSL2_DIST -u root -- sh -c "mountpoint -q /app"
+    wsl -d $WSL2_DIST -u root -- sh -c "mountpoint -q $WSL2_DIST_PATH"
 
     if ($?) {
-      printInfo "$WSL2_DIST /app already mount"
+      printInfo "$WSL2_DIST $WSL2_DIST_PATH already mount"
 
       exit
     }
@@ -1454,7 +1389,7 @@ Example: ./lnmp-docker composer /app/demo install
       $dev_sdx = Get-wsl2_mount_physicaldiskdevice_dev_sdx $MountPhysicalDiskType2WSL2
     }
     else {
-      printInfo "physical disk already mount to $dev_sdx, I will mount $dev_sdx to /app"
+      printInfo "physical disk already mount to $dev_sdx, I will mount $dev_sdx to $WSL2_DIST_PATH"
     }
 
     if (!$dev_sdx) {
@@ -1466,15 +1401,15 @@ Example: ./lnmp-docker composer /app/demo install
     $wsl2_mount_physicaldiskdevice_path = Get-wsl2_mount_physicaldiskdevice_path $MountPhysicalDiskType2WSL2
 
     & $PSScriptRoot/kubernetes/wsl2/bin/wsl2d.ps1 $WSL2_DIST
-    wsl -d $WSL2_DIST -u root -- sh -cx "mkdir -p /app"
-    wsl -d $WSL2_DIST -u root -- sh -cx "mkdir -p $wsl2_mount_physicaldiskdevice_path/app"
-    wsl -d $WSL2_DIST -u root -- sh -cx "mount --bind $wsl2_mount_physicaldiskdevice_path/app /app"
-    wsl -d $WSL2_DIST -u root -- sh -cx "chown 1000:1000 /app"
+    wsl -d $WSL2_DIST -u root -- sh -cx "mkdir -p $WSL2_DIST_PATH"
+    wsl -d $WSL2_DIST -u root -- sh -cx "mkdir -p ${wsl2_mount_physicaldiskdevice_path}${WSL2_DIST_PATH}"
+    wsl -d $WSL2_DIST -u root -- sh -cx "mount --bind ${wsl2_mount_physicaldiskdevice_path}${WSL2_DIST_PATH} $WSL2_DIST_PATH"
+    wsl -d $WSL2_DIST -u root -- sh -cx "chown 1000:1000 $WSL2_DIST_PATH"
   }
 
   "^code$" {
-    if ($env:USE_WSL2_DOCKER_COMPOSE -eq '0') {
-      printError "APP_ROOT is not in WSL2, exit"
+    if ($env:USE_WSL2_BUT_DOCKER_NOT_RUNNING -eq '3') {
+      printError "APP_ROOT is not WSL2 dir"
 
       exit 1
     }
@@ -1486,29 +1421,29 @@ Example: ./lnmp-docker composer /app/demo install
           $path = $path.substring(1)
         }
 
-        wsl -d $WSL2_DIST -u root -- test -d $APP_ROOT/$path
+        wsl -d $WSL2_DIST -u root -- test -d $WSL2_DIST_PATH/$path
         if (!$?) {
           # 不是 dir
-          printError WSL2 $WSL2_DIST [ ${APP_ROOT}/${path} ] `
-            is not`'t a dir`, [ ${APP_ROOT} ] include this dir:
+          printError WSL2 $WSL2_DIST [ $WSL2_DIST_PATH/${path} ] `
+            is not`'t a dir`, [ $WSL2_DIST_PATH ] include this dir:
 
           write-host `n
-          wsl -d $WSL2_DIST -u root -- ls -la $APP_ROOT
+          wsl -d $WSL2_DIST -u root -- ls -la $WSL2_DIST_PATH
 
           write-host `n`n
 
-          exit
+          exit 1
         }
-        printInfo Open WSL2 $WSL2_DIST [ $APP_ROOT/$path ]
-        code --remote wsl+$WSL2_DIST $APP_ROOT/$path
+        printInfo Open WSL2 $WSL2_DIST [ $WSL2_DIST_PATH/$path ]
+        code --remote wsl+$WSL2_DIST $WSL2_DIST_PATH/$path
 
         cd $EXEC_CMD_DIR
 
         exit
       }
     }
-    printInfo Open WSL2 $WSL2_DIST [ $APP_ROOT ]
-    code --remote wsl+$WSL2_DIST $APP_ROOT
+    printInfo Open WSL2 $WSL2_DIST [ $WSL2_DIST_PATH ]
+    code --remote wsl+$WSL2_DIST $WSL2_DIST_PATH
   }
 
   outdated {
@@ -1576,7 +1511,7 @@ Example: ./lnmp-docker composer /app/demo install
       "docker-lnmp.override.yml"
 
     $command, $other = $args
-    & { docker-compose $options $command $other }
+    docker compose $options $command $other
   }
 }